maven build 시 나타나는 오류 해결중

+++ bookforeast/src/main/java/com/project/bookforeast/dto/CodeDTO.java
@@ -17,7 +17,7 @@ import lombok.ToString;
 @Builder
 public class CodeDTO {
 
-	private int codeId;
+	private Long codeId;
 	private String codename;
 	private int upperCodeId;
