+++ bookforeast/src/main/java/com/project/bookforeast/repository/FileGroupRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.project.bookforeast.entity.FileGroup;
 
-public interface FileGroupRepository extends CrudRepository<FileGroup, Integer> {
+public interface FileGroupRepository extends CrudRepository<FileGroup, Long> {
 
 	public FileGroup findFirstByOrderByFilegroupIdDesc();
 }
