+++ bookforeast/src/main/java/com/project/bookforeast/repository/GenreRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.project.bookforeast.entity.LikeGenre;
 
-public interface GenreRepository extends CrudRepository<LikeGenre, Integer> {
+public interface GenreRepository extends CrudRepository<LikeGenre, Long> {
 
 }
